🏗️ [chore] rename & enhance common_commands.go
This commit moves all the common_commands.go to a file named `manuscript_port.go`. This is done because it is more descriptive, domain specific, and accurate to its contents. Additionally, the functions inside have been enhanced in order to better validate and check ports against both the manuscript.yaml and the config file.